+++ apps/twig/src/main/services/environment/local.ts
@@ -1,7 +1,6 @@
-import type { FileWatcherService } from "../file-watcher/service.js";
 import type { FsService } from "../fs/service.js";
 import type { GitService } from "../git/service.js";
-import type { ShellService } from "../shell/service.js";
+import type { ProcessTrackingService } from "../process-tracking/service.js";
 import type { WorkspaceService } from "../workspace/service.js";
 import { BaseEnvironment } from "./base.js";
 import { LocalFilesManager } from "./managers/files.js";
@@ -11,17 +10,14 @@ import { LocalWorkspaceManager } from "./managers/workspace.js";
 import {
   type EnvironmentCapabilities,
   type EnvironmentType,
-  type FilesManager,
   type GitManager,
   LOCAL_CAPABILITIES,
-  type ShellManager,
   type WorkspaceManager,
 } from "./types.js";
 
 export interface LocalEnvironmentDependencies {
-  shellService: ShellService;
+  processTrackingService: ProcessTrackingService;
   workspaceService: WorkspaceService;
-  fileWatcherService: FileWatcherService;
   fsService: FsService;
   gitService: GitService;
 }
@@ -30,16 +26,16 @@ export class LocalEnvironment extends BaseEnvironment {
   readonly type: EnvironmentType = "local";
   readonly capabilities: EnvironmentCapabilities = LOCAL_CAPABILITIES;
 
-  readonly shell: ShellManager;
+  readonly shell: LocalShellManager;
   readonly workspace: WorkspaceManager;
-  readonly files: FilesManager;
+  readonly files: LocalFilesManager;
   readonly git: GitManager;
 
   constructor(deps: LocalEnvironmentDependencies) {
     super();
-    this.shell = new LocalShellManager(deps.shellService);
+    this.shell = new LocalShellManager(deps.processTrackingService);
     this.workspace = new LocalWorkspaceManager(deps.workspaceService);
-    this.files = new LocalFilesManager(deps.fileWatcherService, deps.fsService);
+    this.files = new LocalFilesManager(deps.fsService);
     this.git = new LocalGitManager(deps.gitService);
   }
 }
